fix: typecheck
